pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "databricks-ecommerce-pipeline"
version = "1.0.0"
description = "Production-grade Medallion Architecture pipeline on Databricks"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Gaurav Kumar"},
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "build>=1.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

# ── Ruff Configuration ──
# Ruff = extremely fast Python linter + formatter (replaces flake8 + black + isort)
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort (import sorting)
    "UP",    # pyupgrade
    "B",     # bugbear (common bugs)
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

# ── Pytest Configuration ──
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
